+++ sae-20190506/core/Models/GetWebshellTokenRequest.cs
@@ -1,0 +1,34 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Sae20190506.Models
+{
+    public class GetWebshellTokenRequest : TeaModel {
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>017f39b8-dfa4-4e16-a84b-1dcee4b1****</para>
+        /// </summary>
+        [NameInMap("AppId")]
+        [Validation(Required=false)]
+        public string AppId { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>hello-podsdfsdfsdfsdf</para>
+        /// </summary>
+        [NameInMap("PodName")]
+        [Validation(Required=false)]
+        public string PodName { get; set; }
+
+    }
+
+}