+++ sae-20190506/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("3bf29770-9df4-11ef-bbd9-6d162d1df5f8")]
+[assembly: Guid("691c1580-b6ec-11ef-a51e-93b768b17572")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("1.23.3.0")]
-[assembly: AssemblyFileVersion("1.23.3.0")]
+[assembly: AssemblyVersion("1.23.4.0")]
+[assembly: AssemblyFileVersion("1.23.4.0")]